Split file system libraries into mulitple files
