+++ sqlite3/lib/wasm.dart
@@ -19,5 +19,8 @@ import 'package:meta/meta.dart';
 export 'common.dart' hide CommmonSqlite3;
 
 export 'src/wasm/environment.dart';
-export 'src/wasm/file_system.dart' hide AsynchronousIndexedDbFileSystem;
+export 'src/wasm/file_system.dart' hide debugFileSystem;
+export 'src/wasm/file_system/indexed_db.dart'
+    hide AsynchronousIndexedDbFileSystem;
+export 'src/wasm/file_system/opfs.dart';
 export 'src/wasm/sqlite3.dart';
